remove profiles and identities rather than users
